apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psycopg[binary]>=3.1.0
pydantic-settings>=2.0.0
redis>=5.0
pyjwt>=2.8
passlib[bcrypt]>=1.7
email-validator>=2.0
boto3>=1.34.0
python-multipart>=0.0.9

# Test
pytest>=7.0.0
httpx>=0.25.0
